+++ lib/plugins/index.js
@@ -12,7 +12,7 @@ module.exports = [
   require('./logs.js'),
   require('./metrics.js'),
   require('./print.js'),
-  require('./remove/remove.js'),
+  require('./remove.js'),
   require('./rollback/index.js'),
   require('./slstats/slstats.js'),
   require('./plugin/plugin.js'),
