+++ lib/plugins/remove.js
