+++ test/unit/lib/plugins/remove.test.js
@@ -1,8 +1,8 @@
 'use strict';
 
 const chai = require('chai');
-const Remove = require('../../../../../lib/plugins/remove/remove');
-const Serverless = require('../../../../../lib/Serverless');
+const Remove = require('../../../../lib/plugins/remove');
+const Serverless = require('../../../../lib/Serverless');
 
 const expect = chai.expect;
